requirements.txt
# Parsio Dependencies
# Core GUI Framework
PyQt5>=5.15.0

# AI and LLM Integration
google-generativeai>=0.3.0

# Data Processing (Required for Excel export)
pandas>=1.5.0
openpyxl>=3.0.0

# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0

# Environment Management
python-dotenv>=0.19.0

# Development and Testing (Optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
